refactor(src/styles|templates|utils): JavaScript → TypeScript
