--- src/meta/config.js
@@ -1,22 +1,0 @@
-module.exports = {
-  siteTitle: "ythecombinator",
-  shortSiteTitle: "Full-stack engineer",
-  siteDescription: "Full-stack engineer focused on client-side architecture",
-  siteUrl: "https://ythecombinator.space",
-  pathPrefix: "",
-  siteImage: "preview.jpg",
-  siteLanguage: "en",
-  // author
-  authorName: "Matheus Albuquerque",
-  authorTwitterAccount: "ythecombinator",
-  // info
-  infoTitle: "Matheus Albuquerque",
-  infoTitleNote: "It's me, Matheus.",
-  // manifest.json
-  manifestName: "ythecombinator | Full-Stack Engineer",
-  manifestShortName: "Matheus",
-  manifestStartUrl: "/",
-  manifestBackgroundColor: "white",
-  manifestThemeColor: "#5b737d",
-  manifestDisplay: "standalone"
-};
