+++ src/utils/array.ts
@@ -1,4 +1,4 @@
-const getRandomItem = items => {
+const getRandomItem = <T>(items: T[]) => {
   return items[Math.floor(Math.random() * items.length)];
 };
