+++ src/utils/date.ts
@@ -1,0 +1,8 @@
+import dayjs from 'dayjs';
+import advancedFormat from 'dayjs/plugin/advancedFormat';
+
+dayjs.extend(advancedFormat);
+
+const format = (date: string) => dayjs(date).format("MMM Do YYYY");
+
+export { format };
