+++ src/utils/number.ts
@@ -1,8 +1,8 @@
-const getRandomInt = (min, max) => {
+const getRandomInt = (min: number, max: number) => {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 };
 
-const getRandomFloat = (min, max) => {
+const getRandomFloat = (min: number, max: number) => {
   return (Math.random() * (min - max) + min).toFixed(1);
 };
